Move EditorChildrenProps to types file
