+++ Toolbar/LinkButton.tsx
@@ -2,7 +2,7 @@ import icons from '@rmccue/sfsymbols';
 import React from 'react';
 
 import Button from './Button';
-import { EditorChildrenProps } from '../Editor';
+import { EditorChildrenProps } from '../types';
 
 export default function FormatButton( props: EditorChildrenProps ) {
 	return (
